Great bugfix in module attach/detach scheme.
Shadow module pointers are introduced to be used on module detach.
Since detach function is called from IModuleBase destructor
(at which time the module itself is already destroyed, and so is
its vptr), comparison of pointer to IModuleBase passed to Detach()
with any qualified pointer to module (say, IReader*) -
which is already invalid - will NOT succeed.

So we keep "shadow" pointers to modules' bases (IModuleBase*)
along with original pointers to modules.