Re-implement displacement and immediate value support.
Use templates to allow bit-perfect value copying.

This is not perfect, as various types (like enums) which can be passed
to Insn::AddImmediate()/Insn::AddDisplacement() may have different
runtime sizes, and system will use exactly that size.

It would be good to allow a user to directly specify the desired
imm/disp operand size and then perform an integer cast (either unsigned
or signed), but that is going to be too difficult and over-engineered.