Make string literals usable again.
1)
Do not bail out in linker when a bytepool reference is to be
auto-placed.
Actually, placing bytepool references is possible as we have the proper
limit field; it is unclear why this was prohibited earlier.

2)
If a string literal is specified in-place (without symbol),
it was not assigned any address since linker does not handle such
references at all. Fix it by introducing a new linker direct link
interface method which handles a reference (not symbol).