Add recursive dep handling in production mode.
This adds a new file to CMake inter-project API:
"CMakeConfigureProduction.cmake".

It shall contain calls to
- add_production_build_library (...)
- add_production_build_library_custom (...)
- list (APPEND LIBRARIES ...)

Its purpose is to recursively gather base include pathes for all
sub-projects and add them to the top-level project.

It is needed as API headers for sub-projects often reference
API headers of their own sub-projects.